20241219 i can now move the player onto the vehicle and press shift, this puts the player in the vehicle, now when the player moves the vehicle moves too, press shift again to get out then the player can move and the vehicle will stay still, tested with a second player and a second vehicle i learned something new about dynamic property access, i created a print method in the mapData class to help show a text representation of the various variables of mapData.tiles[]
TODO
players should not be able to occupy the same space as other players, ever.
vehicles  should not be able to occupy the same space as other vehicles, ever.
doing so would mess up the mapData.tiles[].playerIndex / vehicleIndex.
players can occupy the same space as a vehicle but only if it is empty

i also need to make the vehicle interact with the walls and rubble etc