20250114 made 2 creatures , made them walk towards a goal, they should meet in the middle and get stuck on each other also made collisions between creatures and - terrain, players, vehicles, other creatures made 2 continuous rows of creatures, one going east one going west, you can notice how one moves normally as the the leader has no creature in it's way, and the other turns from contiuous to a sort of dotted line of creatures, as the leader is at the back and the other creatures are in it's way - this is because i am updating the creatures movement in order of the index number of the creatures, which is arbitrary, when i implement the priority array properly we will be updating the creatures that intend on moving east in order of the eastern most creature, so the others will not have a creature in their way
now i want to implement the more complex movement rules, like the priority arrays
i will create a priority array class, that we pass a reference of in to each creature, i've created methods to loop through the priority arrays and to change the priority, which seems to work ok

if i add an index of a creature, that wants to travel north, to the north array, then i want to sort the array, or more efficiently i want to insert the new index in the correct order.
the correct order will be northern most creatures first, (for the north array)
since the north array should be sorted from the previous update, i can make use of that to efficiently add creatures that still want to go north to the north array, creatures that were in a different array previously - east, west, stationary or south, but now want to go north will be more complex to insert into the north array

often times i will process the creatures that want to go north in an order that was already largely sorted in order of most northern, because creatures that are travelling north most likely will go north again, so assuming this is true more often than not, we should use an insert sort.

so for example, i have a list of creature indexes that travelled north previously, i go through one by one, starting with the most northern, if it wants to go north again, i will add it to the now empty north array, the next one i add will be either equally northern to the ones already in the array or less northern, so i will check the last item in the north array, if this item is less northern than that (which it should be) then we add it to the end of the array, if not check the second to last item etc. until we find the correct place to insert it -
this only works for the north array and only if it was already empty. the east, west and stationary arrays will be sorted in ways irrelevant to northern-ness, the south array will be ordered in the opposite way to the north array, with southern most creatures appearing first in the south array, so we could maybe process the north array and south array at the same time, the south array in reverse order

as for the east, west and stationary arrays, these will have to be sorted in a less efficient way

so far i have 2 actors, one in the east that wants to move west and one in the west that wants to move east, they meet in the middle and are stopped by each other
after pathfinding, they are added to priority arrays based on their intended direction of travel, at the end of the movement update these arrays are saved into an 'old' version of the array and the arrays are cleared

TODO
instead of going through the array of creatures, in the order of their index numbers, i want to change it so that i go through based on the priority array,
i think i need to add the relevant x or y coord as well as the index to the priority array, to aid sorting
i need to implement the sorting, if i do these correctly the 2 continuous rows of creatures should stay continuous