20250206 pathfinding, now i will have the creature remember its 'tail' or the space it came from this works, although in my map the creature does still get stuck , it gets to a dead end and goes back on its tail, it can then either move south or east, it goes east, which is the way it have previously came from it ends up in another dead end and comes back to the first dead end and run repeats, but at least it is going back and forth over several tiles rather than just flicking between one tile and another
to try and help the above, perhaps i can create landmarks, these are placed on the map by a creature for all creatures to see, warning them not to go in that direction. if you travel back on your own tail that means you hit a dead end. a simplified version of this landmark system can be to place a 'deadend' landmark any time you have to travel to your own tail, place it on your position and now that dead end is one tile shorter. if numerous creatures enter a long corridor that is a dead end, they will eventually fill it up with dead ends, and no creatures will bother going that way anymore. a creature will decide to travel over its own tail when the 3 other directions were not viable, that can be because they are paths, or dead ends, if the creature turned back because it was a dead end landmark, and it wants to make it's own landmark, then delete the one we encoutnered as we have set a more recent one - this doesn't really work, first of all, different creatures might be looking for different things, so it might not be a deadend to one creature, it might be exactly where it was trying to go. also if at the end of a long deadend corridor, there is space to turn around then it would not be marked as a dead end.

todo i should resume with my plan to do a wall rider
i need to think about the distinction between creatures that are exploring and creatures who are trying to reach a defined goal position, these creatures should navigate in different ways. landmarks could be useful, but only at intersections - at a cross roads these landmarks or signposts could be useful