20250705 once the creatures gather all resources they don't explore the rest of the map, the go up to bits they have explored before, but no further. one creature index number 5, eventually breaks out and explores some but it loops around back to explored territory, but as it's dead end flag is set to true it ends up stuck going back and forth between 2 tiles as they are the lowest explored numbers. maybe when dead end is true, i should only move to the lowest explored neighbour if it is lower than the creature's tile's explored number, if we are on the lowest already then maybe set dead end to false? no i think something else is wrong here, it gets to a explored number that is 6, then there is no lower number, how is that possible? the explored number that should be 5 is 27, when did that happen? i was supposed to only update the explored number if we have a new lower number, fixed that - there was a typo, this does make the behaviour different.
the way the resource markers work is not right, i think we should only be putting these down when we are following a path back to the base, it seems that creature 5 gets to the resource when it is depleted, then keeps going past the resource laying resource markers?  i think, it needs more testing. i think this is why creature 5 is dancing around the depleted resource, it's running a counter down to 0 i think which is not the intended behaviour